requirements.txt
fastapi[standard]
pydantic-settings
alembic
sqlmodel
asyncmy
passlib[bcrypt]
bcrypt==3.2.2
python-jose[cryptography]
python-dotenv
requests